Split serializer tests into EMLSerializerTests.swift
After rebasing onto main's byte-level parser tests, EMLTests.swift
exceeded SwiftLint's 400-line file_length limit. Move the EML Serializer
suite into its own file.
